+++ simple-stack/src/main/java/com/zhuinden/simplestack/BackstackDelegate.java
@@ -433,6 +433,17 @@ public class BackstackDelegate {
     }
 
     /**
+     * Returns if a given scope exists.
+     *
+     * @param scopeTag   the scope tag
+     *
+     * @return whether the scope exists
+     */
+    public boolean hasScope(@NonNull String scopeTag) {
+        return getManager().hasScope(scopeTag);
+    }
+
+    /**
      * Returns if a service is bound to the scope of the {@link ScopeKey} by the provided tag.
      *
      * @param scopeKey   the scope key
@@ -454,6 +465,18 @@ public class BackstackDelegate {
     @NonNull
     public <T> T getService(@NonNull ScopeKey scopeKey, @NonNull String serviceTag) {
         return getManager().getService(scopeKey, serviceTag);
+    }
+
+    /**
+     * Attempts to look-up the service in all currently existing scopes, starting from the last added scope.
+     * Returns whether the service exists in any scopes.
+     *
+     * @param serviceTag the tag of the service
+     * @return whether the service exists in any active scopes
+     * @throws IllegalStateException if the service doesn't exist in any scope
+     */
+    public boolean canFindService(@NonNull String serviceTag) {
+        return getManager().canFindService(serviceTag);
     }
 
     /**